+++ server/prisma/migrations/20260129091635_123/migration.sql
@@ -1,0 +1,30 @@
+-- CreateTable
+CREATE TABLE "User" (
+    "id" INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
+    "botId" TEXT NOT NULL,
+    "userId" TEXT NOT NULL,
+    "username" TEXT NOT NULL,
+    "name" TEXT,
+    "createdAt" DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    "blocked" BOOLEAN NOT NULL DEFAULT false
+);
+
+-- CreateTable
+CREATE TABLE "Message" (
+    "id" INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
+    "botId" TEXT NOT NULL,
+    "roomId" TEXT NOT NULL,
+    "userId" TEXT NOT NULL,
+    "messageType" TEXT NOT NULL,
+    "text" TEXT NOT NULL,
+    "attachments" JSONB,
+    "meta" JSONB,
+    "createdAt" DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    CONSTRAINT "Message_userId_botId_fkey" FOREIGN KEY ("userId", "botId") REFERENCES "User" ("userId", "botId") ON DELETE RESTRICT ON UPDATE CASCADE
+);
+
+-- CreateIndex
+CREATE UNIQUE INDEX "User_botId_userId_key" ON "User"("botId", "userId");
+
+-- CreateIndex
+CREATE INDEX "Message_botId_roomId_idx" ON "Message"("botId", "roomId");